+++ ThirdPartyAdapters/pangle/pangle/src/main/java/com/google/ads/mediation/pangle/PangleMediationAdapter.java
@@ -15,6 +15,7 @@
 package com.google.ads.mediation.pangle;
 
 import static com.google.ads.mediation.pangle.PangleConstants.ERROR_INVALID_SERVER_PARAMETERS;
+import static com.google.ads.mediation.pangle.PangleConstants.isChildUser;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -22,9 +23,11 @@ import android.text.TextUtils;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
-import com.bytedance.sdk.openadsdk.api.PAGConstant.PAGDoNotSellType;
+import com.bytedance.sdk.openadsdk.api.PAGConstant;
 import com.bytedance.sdk.openadsdk.api.PAGConstant.PAGGDPRConsentType;
+import com.bytedance.sdk.openadsdk.api.bidding.PAGBiddingRequest;
 import com.bytedance.sdk.openadsdk.api.init.BiddingTokenCallback;
+import com.bytedance.sdk.openadsdk.api.init.PAGConfig;
 import com.google.ads.mediation.pangle.PangleInitializer.Listener;
 import com.google.ads.mediation.pangle.renderer.PangleAppOpenAd;
 import com.google.ads.mediation.pangle.renderer.PangleBannerAd;
@@ -32,7 +35,6 @@ import com.google.ads.mediation.pangle.renderer.PangleInterstitialAd;
 import com.google.ads.mediation.pangle.renderer.PangleNativeAd;
 import com.google.ads.mediation.pangle.renderer.PangleRewardedAd;
 import com.google.android.gms.ads.AdError;
-import com.google.android.gms.ads.MobileAds;
 import com.google.android.gms.ads.VersionInfo;
 import com.google.android.gms.ads.mediation.InitializationCompleteCallback;
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback;
@@ -68,7 +70,7 @@ public class PangleMediationAdapter extends RtbAdapter {
   private final PangleInitializer pangleInitializer;
   private final PangleSdkWrapper pangleSdkWrapper;
   private final PangleFactory pangleFactory;
-  private final PanglePrivacyConfig panglePrivacyConfig;
+
 
   private PangleAppOpenAd appOpenAd;
   private PangleBannerAd bannerAd;
@@ -77,41 +79,47 @@ public class PangleMediationAdapter extends RtbAdapter {
   private PangleRewardedAd rewardedAd;
 
   private static int gdpr = -1;
-  private static int ccpa = -1;
 
   public PangleMediationAdapter() {
     pangleInitializer = PangleInitializer.getInstance();
     pangleSdkWrapper = new PangleSdkWrapper();
     pangleFactory = new PangleFactory();
-    panglePrivacyConfig = new PanglePrivacyConfig(pangleSdkWrapper);
+
   }
 
   @VisibleForTesting
   PangleMediationAdapter(
       PangleInitializer pangleInitializer,
       PangleSdkWrapper pangleSdkWrapper,
-      PangleFactory pangleFactory,
-      PanglePrivacyConfig panglePrivacyConfig) {
+      PangleFactory pangleFactory) {
     this.pangleInitializer = pangleInitializer;
     this.pangleSdkWrapper = pangleSdkWrapper;
     this.pangleFactory = pangleFactory;
-    this.panglePrivacyConfig = panglePrivacyConfig;
   }
 
   @Override
   public void collectSignals(
       @NonNull RtbSignalData rtbSignalData, @NonNull SignalCallbacks signalCallbacks) {
+    if (isChildUser()) {
+      signalCallbacks.onFailure(PangleConstants.createChildUserError());
+      return;
+    }
     // The user data needs to be set for it to be included in the signals.
     Bundle networkExtras = rtbSignalData.getNetworkExtras();
     if (networkExtras != null && networkExtras.containsKey(PangleExtras.Keys.USER_DATA)) {
       pangleSdkWrapper.setUserData(networkExtras.getString(PangleExtras.Keys.USER_DATA, ""));
     }
-    pangleSdkWrapper.getBiddingToken(new BiddingTokenCallback() {
-      @Override
-      public void onBiddingTokenCollected(String biddingToken) {
-        signalCallbacks.onSuccess(biddingToken);
-      }
-    });
+    PAGBiddingRequest biddingRequest = new PAGBiddingRequest();
+    biddingRequest.setAdxId(PangleConstants.ADX_ID);
+    pangleSdkWrapper.getBiddingToken(
+        rtbSignalData.getContext(),
+        biddingRequest,
+        new BiddingTokenCallback() {
+          @Override
+          public void onBiddingTokenCollected(String biddingToken) {
+            signalCallbacks.onSuccess(biddingToken);
+          }
+        });
   }
 
   @Override
@@ -119,6 +127,10 @@ public class PangleMediationAdapter extends RtbAdapter {
       @NonNull Context context,
       @NonNull final InitializationCompleteCallback initializationCompleteCallback,
       @NonNull List<MediationConfiguration> list) {
+    if (isChildUser()) {
+      initializationCompleteCallback.onInitializationFailed(PangleConstants.ERROR_MSG_CHILD_USER);
+      return;
+    }
     HashSet<String> appIds = new HashSet<>();
     for (MediationConfiguration mediationConfiguration : list) {
       Bundle serverParameters = mediationConfiguration.getServerParameters();
@@ -146,8 +158,6 @@ public class PangleMediationAdapter extends RtbAdapter {
       Log.w(TAG, message);
     }
 
-    panglePrivacyConfig.setCoppa(
-        MobileAds.getRequestConfiguration().getTagForChildDirectedTreatment());
     pangleInitializer.initialize(
         context,
         appId,
@@ -221,9 +231,13 @@ public class PangleMediationAdapter extends RtbAdapter {
   public void loadAppOpenAd(
       @NonNull MediationAppOpenAdConfiguration adConfiguration,
       @NonNull MediationAdLoadCallback<MediationAppOpenAd, MediationAppOpenAdCallback> callback) {
+    if (isChildUser()) {
+      callback.onFailure(PangleConstants.createChildUserError());
+      return;
+    }
     appOpenAd =
         pangleFactory.createPangleAppOpenAd(
-            adConfiguration, callback, pangleInitializer, pangleSdkWrapper, panglePrivacyConfig);
+            adConfiguration, callback, pangleInitializer, pangleSdkWrapper);
     appOpenAd.render();
   }
 
@@ -231,9 +245,14 @@ public class PangleMediationAdapter extends RtbAdapter {
   public void loadBannerAd(
       @NonNull MediationBannerAdConfiguration adConfiguration,
       @NonNull MediationAdLoadCallback<MediationBannerAd, MediationBannerAdCallback> callback) {
+    if (isChildUser()) {
+      callback.onFailure(PangleConstants.createChildUserError());
+      return;
+    }
+
     bannerAd =
         pangleFactory.createPangleBannerAd(
-            adConfiguration, callback, pangleInitializer, pangleSdkWrapper, panglePrivacyConfig);
+            adConfiguration, callback, pangleInitializer, pangleSdkWrapper);
     bannerAd.render();
   }
 
@@ -243,9 +262,14 @@ public class PangleMediationAdapter extends RtbAdapter {
       @NonNull
           MediationAdLoadCallback<MediationInterstitialAd, MediationInterstitialAdCallback>
               callback) {
+    if (isChildUser()) {
+      callback.onFailure(PangleConstants.createChildUserError());
+      return;
+    }
+
     interstitialAd =
         pangleFactory.createPangleInterstitialAd(
-            adConfiguration, callback, pangleInitializer, pangleSdkWrapper, panglePrivacyConfig);
+            adConfiguration, callback, pangleInitializer, pangleSdkWrapper);
     interstitialAd.render();
   }
 
@@ -253,9 +277,14 @@ public class PangleMediationAdapter extends RtbAdapter {
   public void loadNativeAd(
       @NonNull MediationNativeAdConfiguration adConfiguration,
       @NonNull MediationAdLoadCallback<UnifiedNativeAdMapper, MediationNativeAdCallback> callback) {
+    if (isChildUser()) {
+      callback.onFailure(PangleConstants.createChildUserError());
+      return;
+    }
+
     nativeAd =
         pangleFactory.createPangleNativeAd(
-            adConfiguration, callback, pangleInitializer, pangleSdkWrapper, panglePrivacyConfig);
+            adConfiguration, callback, pangleInitializer, pangleSdkWrapper);
     nativeAd.render();
   }
 
@@ -263,9 +292,13 @@ public class PangleMediationAdapter extends RtbAdapter {
   public void loadRewardedAd(
       @NonNull MediationRewardedAdConfiguration adConfiguration,
       @NonNull MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback> callback) {
+    if (isChildUser()) {
+      callback.onFailure(PangleConstants.createChildUserError());
+      return;
+    }
     rewardedAd =
         pangleFactory.createPangleRewardedAd(
-            adConfiguration, callback, pangleInitializer, pangleSdkWrapper, panglePrivacyConfig);
+            adConfiguration, callback, pangleInitializer, pangleSdkWrapper);
     rewardedAd.render();
   }
 
@@ -290,7 +323,7 @@ public class PangleMediationAdapter extends RtbAdapter {
       Log.w(TAG, "Invalid GDPR value. Pangle SDK only accepts -1, 0 or 1.");
       return;
     }
-    if (pangleSdkWrapper.isInitSuccess()) {
+    if (pangleSdkWrapper.isInitSuccess() && !isChildUser()) {
       pangleSdkWrapper.setGdprConsent(gdpr);
     }
     PangleMediationAdapter.gdpr = gdpr;
@@ -301,33 +334,31 @@ public class PangleMediationAdapter extends RtbAdapter {
   }
 
   /**
-   * Set the CCPA setting in Pangle SDK.
+   * Set the PA setting in Pangle SDK.
    *
-   * @param ccpa an {@code Integer} value that indicates whether the user opts in of the "sale" of
-   *     the "personal information" under CCPA. See <a
+   * @param pa an {@code Integer} value that Indicates whether the user agrees to the delivery of
+   *     personalized ads. If not passed, it is assumed to be agreed. See <a
    *     href="https://www.pangleglobal.com/integration/android-initialize-pangle-sdk">Pangle's
    *     documentation</a> for more information about what values may be provided.
    */
-  public static void setDoNotSell(@PAGDoNotSellType int ccpa) {
-    setDoNotSell(ccpa, new PangleSdkWrapper());
+  public static void setPAConsent(@PAGConstant.PAGPAConsentType int pa) {
+    setPAConsent(pa, new PangleSdkWrapper());
   }
 
   @VisibleForTesting
-  static void setDoNotSell(@PAGDoNotSellType int ccpa, PangleSdkWrapper pangleSdkWrapper) {
-    if (ccpa != PAGDoNotSellType.PAG_DO_NOT_SELL_TYPE_SELL
-        && ccpa != PAGDoNotSellType.PAG_DO_NOT_SELL_TYPE_NOT_SELL
-        && ccpa != PAGDoNotSellType.PAG_DO_NOT_SELL_TYPE_DEFAULT) {
+  static void setPAConsent(
+      @PAGConstant.PAGPAConsentType int pa, PangleSdkWrapper pangleSdkWrapper) {
+    if (pa != PAGConstant.PAGPAConsentType.PAG_PA_CONSENT_TYPE_CONSENT
+        && pa != PAGConstant.PAGPAConsentType.PAG_PA_CONSENT_TYPE_NO_CONSENT) {
       // no-op
-      Log.w(TAG, "Invalid CCPA value. Pangle SDK only accepts -1, 0 or 1.");
+      Log.w(TAG, "Invalid PA value. Pangle SDK only accepts 0 or 1.");
       return;
     }
-    if (pangleSdkWrapper.isInitSuccess()) {
-      pangleSdkWrapper.setDoNotSell(ccpa);
-    }
-    PangleMediationAdapter.ccpa = ccpa;
+    pangleSdkWrapper.setPAConsent(pa);
   }
 
-  public static int getDoNotSell() {
-    return ccpa;
+  public static int getPAConsent() {
+    return PAGConfig.getPAConsent();
   }
+
 }