+++ ThirdPartyAdapters/pangle/pangle/src/main/java/com/google/ads/mediation/pangle/PangleSdkWrapper.java
@@ -16,12 +16,12 @@ package com.google.ads.mediation.pangle;
 
 import android.content.Context;
 import androidx.annotation.NonNull;
-import com.bytedance.sdk.openadsdk.api.PAGConstant.PAGChildDirectedType;
-import com.bytedance.sdk.openadsdk.api.PAGConstant.PAGDoNotSellType;
+import com.bytedance.sdk.openadsdk.api.PAGConstant;
 import com.bytedance.sdk.openadsdk.api.PAGConstant.PAGGDPRConsentType;
 import com.bytedance.sdk.openadsdk.api.banner.PAGBannerAd;
 import com.bytedance.sdk.openadsdk.api.banner.PAGBannerAdLoadListener;
 import com.bytedance.sdk.openadsdk.api.banner.PAGBannerRequest;
+import com.bytedance.sdk.openadsdk.api.bidding.PAGBiddingRequest;
 import com.bytedance.sdk.openadsdk.api.init.BiddingTokenCallback;
 import com.bytedance.sdk.openadsdk.api.init.PAGConfig;
 import com.bytedance.sdk.openadsdk.api.init.PAGSdk;
@@ -54,24 +54,24 @@ public class PangleSdkWrapper {
     return PAGSdk.isInitSuccess();
   }
 
-  void setChildDirected(@PAGChildDirectedType int childDirectedType) {
-    PAGConfig.setChildDirected(childDirectedType);
-  }
 
   void setGdprConsent(@PAGGDPRConsentType int gdpr) {
     PAGConfig.setGDPRConsent(gdpr);
   }
 
-  void setDoNotSell(@PAGDoNotSellType int ccpa) {
-    PAGConfig.setDoNotSell(ccpa);
+  void setPAConsent(@PAGConstant.PAGPAConsentType int pa) {
+    PAGConfig.setPAConsent(pa);
   }
 
   void setUserData(String userData) {
     PAGConfig.setUserData(userData);
   }
 
-  void getBiddingToken(BiddingTokenCallback biddingTokenCallback) {
-    PAGSdk.getBiddingToken(biddingTokenCallback);
+  void getBiddingToken(
+      Context context,
+      PAGBiddingRequest biddingRequest,
+      BiddingTokenCallback biddingTokenCallback) {
+    PAGSdk.getBiddingToken(context, biddingRequest, biddingTokenCallback);
   }
 
   String getSdkVersion() {